Malloc Free

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -12,15 +12,14 @@ int **g;
 int i, j = 0;
 if (width <= 0 || height <= 0)
 return (NULL);
-
-g = (int) malloc(height * sizeof(*g));
+g = malloc(height * sizeof(*g));
 
 if (g == NULL)
 return (NULL);
 
 for (i = 0; i < height; i++)
 {
-g[i] = (int) malloc(width * sizeof(**g));
+g[i] = malloc(width * sizeof(**g));
 if (g[i] == NULL)
 {
 free(g);
